+++ CHANGELOG.md
@@ -20,6 +20,7 @@ omnibus-gitlab repository.
 - Upgrade Docker registry to 2.7.1 !2970
 - Upgrade Nginx to 1.12.2 !2988
 - Bump gitlab-elasticsearch-indexer to 1.0.0
+- Include experimental Docker Distribution Pruner !2946
 
 11.7.0